refactor(tui): split binding-stack reachability scanning into keymap-reachability.ts
keymap-dispatch.ts was pinned at 499 lines with only one line of headroom.
Move the reachability helpers (prefixReachable, inputPassthroughReachable,
scanReachability, bindingReachability) and the BindingReachability type into
a dedicated keymap-reachability.ts module.

The dispatchKeyEvent hot path keeps its early-exit helpers unchanged; the
perf-budget test still reads one config on a top-group hit. Cold callers
(armPrefixNow, bindingReachability) now import the unified scanner.

keymap-dispatch.ts drops from 499 to 429 lines.
