+++ packages/kobe/src/tui-react/component/help-dialog.tsx
@@ -19,7 +19,8 @@ import { type ScrollBoxRenderable, TextAttributes } from "@opentui/core"
 import { useMemo, useRef } from "react"
 import { formatChord } from "../../tui/lib/chord-glyphs"
 import { type HelpGrammarSection, type HelpSurface, grammarHelpSections } from "../../tui/lib/help-groups"
-import { type BindingReachability, currentPrefixConfiguration } from "../../tui/lib/keymap-dispatch"
+import { currentPrefixConfiguration } from "../../tui/lib/keymap-dispatch"
+import type { BindingReachability } from "../../tui/lib/keymap-reachability"
 import { KobeKeymap, useKeymapVersion } from "../context/keybindings"
 import { useTheme } from "../context/theme"
 import { tKeys, useT } from "../i18n"
